# docs/api/swagger.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Skyline API",
        "version": "0.1.0"
    },
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login",
                "description": "Login & get user profile.",
                "operationId": "login_api_v1_login_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Credential"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedMessage"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/profile": {
            "get": {
                "tags": [
                    "Login"
                ],
                "summary": "Get Profile",
                "description": "Get user profile.",
                "operationId": "get_profile_api_v1_profile_get",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedMessage"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/switch_project/{project_id}": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Switch Project",
                "description": "Switch project.",
                "operationId": "switch_project_api_v1_switch_project__project_id__post",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Project Id",
                            "type": "string"
                        },
                        "name": "project_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedMessage"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/policies": {
            "get": {
                "tags": [
                    "Policy"
                ],
                "summary": "List Policies",
                "description": "List policies and permissions",
                "operationId": "list_policies_api_v1_policies_get",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Policies"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedMessage"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InternalServerErrorMessage"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/policies/check": {
            "post": {
                "tags": [
                    "Policy"
                ],
                "summary": "Check Policies",
                "description": "Check policies permissions",
                "operationId": "check_policies_api_v1_policies_check_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PoliciesRules"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Policies"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedMessage"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenMessage"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InternalServerErrorMessage"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Credential": {
                "title": "Credential",
                "required": [
                    "region",
                    "domain",
                    "username",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "region": {
                        "title": "Region",
                        "type": "string"
                    },
                    "domain": {
                        "title": "Domain",
                        "type": "string"
                    },
                    "username": {
                        "title": "Username",
                        "type": "string"
                    },
                    "password": {
                        "title": "Password",
                        "type": "string"
                    }
                },
                "example": {
                    "region": "RegionOne",
                    "username": "admin",
                    "domain": "default",
                    "password": "admin"
                }
            },
            "Domain": {
                "title": "Domain",
                "required": [
                    "id",
                    "name"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "string"
                    },
                    "name": {
                        "title": "Name",
                        "type": "string"
                    }
                }
            },
...
            "ForbiddenMessage": {
                "title": "ForbiddenMessage",
                "required": [
                    "detail"
                ],
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "string"
                    }
                }
            },
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "InternalServerErrorMessage": {
                "title": "InternalServerErrorMessage",
                "required": [
                    "detail"
                ],
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "string"
                    }
                }
            },
            "License": {
                "title": "License",
                "required": [
                    "name",
                    "summary",
                    "macs",
                    "features",
                    "start",
                    "end"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "title": "Name",
                        "type": "string"
                    },
                    "summary": {
                        "title": "Summary",
                        "type": "string"
                    },
                    "macs": {
                        "title": "Macs",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "features": {
                        "title": "Features",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "start": {
                        "title": "Start",
                        "type": "string"
                    },
                    "end": {
                        "title": "End",
                        "type": "string"
                    }
                }
            },
...
            "Policies": {
                "title": "Policies",
                "required": [
                    "policies"
                ],
                "type": "object",
                "properties": {
                    "policies": {
                        "title": "Policies",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Policy"
                        }
                    }
                }
            },
            "PoliciesRules": {
                "title": "PoliciesRules",
                "required": [
                    "rules"
                ],
                "type": "object",
                "properties": {
                    "rules": {
                        "title": "Rules",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "Policy": {
                "title": "Policy",
                "required": [
                    "rule",
                    "allowed"
                ],
                "type": "object",
                "properties": {
                    "rule": {
                        "title": "Rule",
                        "type": "string"
                    },
                    "allowed": {
                        "title": "Allowed",
                        "type": "boolean"
                    }
                }
            },
            "Profile": {
                "title": "Profile",
                "required": [
                    "keystone_token",
                    "region",
                    "project",
                    "user",
                    "roles",
                    "keystone_token_exp",
                    "exp",
                    "uuid",
                    "version"
                ],
                "type": "object",
                "properties": {
                    "keystone_token": {
                        "title": "Keystone Token",
                        "type": "string"
                    },
                    "region": {
                        "title": "Region",
                        "type": "string"
                    },
                    "project": {
                        "$ref": "#/components/schemas/Project"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "roles": {
                        "title": "Roles",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    },
                    "keystone_token_exp": {
                        "title": "Keystone Token Exp",
                        "type": "string"
                    },
                    "base_roles": {
                        "title": "Base Roles",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "base_domains": {
                        "title": "Base Domains",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "endpoints": {
                        "title": "Endpoints",
                        "type": "object"
                    },
                    "projects": {
                        "title": "Projects",
                        "type": "object"
                    },
                    "exp": {
                        "title": "Exp",
                        "type": "integer"
                    },
                    "uuid": {
                        "title": "Uuid",
                        "type": "string"
                    },
                    "version": {
                        "title": "Version",
                        "type": "string"
                    },
                    "license": {
                        "$ref": "#/components/schemas/License"
                    },
                    "currency": {
                        "title": "Currency",
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                }
            },
            "Project": {
                "title": "Project",
                "required": [
                    "id",
                    "name",
                    "domain"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "string"
                    },
                    "name": {
                        "title": "Name",
                        "type": "string"
                    },
                    "domain": {
                        "$ref": "#/components/schemas/Domain"
                    }
                }
            },
            "Role": {
                "title": "Role",
                "required": [
                    "id",
                    "name"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "string"
                    },
                    "name": {
                        "title": "Name",
                        "type": "string"
                    }
                }
            },
...
            "UnauthorizedMessage": {
                "title": "UnauthorizedMessage",
                "required": [
                    "detail"
                ],
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "string"
                    }
                }
            },
...
            "User": {
                "title": "User",
                "required": [
                    "id",
                    "name",
                    "domain"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "string"
                    },
                    "name": {
                        "title": "Name",
                        "type": "string"
                    },
                    "domain": {
                        "$ref": "#/components/schemas/Domain"
                    }
                }
            },
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            },
...
        }
    }
}